% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_mean_surface.R
\name{compute_mean_surface}
\alias{compute_mean_surface}
\title{Compute Interpolated Mean Surface for Spatiotemporal Model}
\usage{
compute_mean_surface(results, data, sites, X, Z, DNO, CovXNO, CovZNO, tau)
}
\arguments{
\item{results}{A list containing the output from \code{STModelWeibullMCMC}, including:
\itemize{
\item \code{MW}: Samples for parameter W.
\item \code{MWT}: Acceptance indicators for parameter W.
\item \code{MMj}: Samples for parameter M.
\item \code{MMT}: Acceptance indicators for parameter M.
\item \code{Mvw}: Samples for parameter sigma^2_w.
\item \code{Mvm}: Samples for parameter sigma^2_m.
\item \code{MBeta}: Samples for parameter Beta.
\item \code{MbwT}: Acceptance indicators for phi_w.
\item \code{MbmT}: Acceptance indicators for phi_m.
\item \code{MPsi}: Samples for parameter Psi.
}}

\item{data}{A matrix of occurrence times of the event at monitoring stations.}

\item{sites}{A matrix with geographic coordinates of the monitoring stations.}

\item{X}{Covariates for the scale parameter of the Weibull intensity.}

\item{Z}{Covariates for the shape parameter of the Weibull intensity.}

\item{DNO}{A grid of points where interpolation is to be performed.}

\item{CovXNO}{Covariates for the scale parameter at the grid points.}

\item{CovZNO}{Covariates for the shape parameter at the grid points.}

\item{tau}{A vector of temporal points for which the mean surface is computed.}
}
\value{
A matrix (\code{Surface}) containing the interpolated mean values at the grid points
for each temporal point in \code{tau}. The first column contains the mean values at the initial
time step, and subsequent columns contain the differences between consecutive time steps.
}
\description{
This function interpolates the mean values on a grid of points (\code{DNO}) for a spatiotemporal
nonhomogeneous Poisson model with Weibull intensity. The function utilizes the MCMC outputs
from \code{STModelWeibullMCMC} and applies a Gaussian process-based interpolation.
}
