% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_mean_surfaceGOEL.R
\name{compute_mean_surfaceGOEL}
\alias{compute_mean_surfaceGOEL}
\title{Compute Interpolated Mean Surface for Spatiotemporal Model}
\usage{
compute_mean_surfaceGOEL(results, sites, X, Z, DNO, CovXNO, CovZNO, tau)
}
\arguments{
\item{results}{A list containing the output from \code{STModelGoelMCMC}, including:
\itemize{
\item \code{Mgama}: MCMC chain for the parameter gamma.
\item \code{Meta}: MCMC chain for the parameter eta.
\item \code{Mv}: MCMC chain for the parameter sigma^2.
\item \code{Mb}: MCMC chain for the parameter phi.
\item \code{MW}: MCMC chain for the parameter W.
\item \code{MPsi}: MCMC chain for the parameter Psi.
}}

\item{sites}{A matrix with geographic coordinates of the monitoring stations.}

\item{X}{Covariates for the scale parameter of the Weibull intensity.}

\item{Z}{Covariates for the shape parameter of the Weibull intensity.}

\item{DNO}{A grid of points where interpolation is to be performed.}

\item{CovXNO}{Covariates for the scale parameter at the grid points.}

\item{CovZNO}{Covariates for the shape parameter at the grid points.}

\item{tau}{A vector of temporal points for which the mean surface is computed.}
}
\value{
A list: (\code{Surface}) containing the interpolated mean values at the grid points
for each temporal point in \code{tau}. The first column contains the mean values at the initial
time step, and subsequent columns contain the differences between consecutive time steps.
}
\description{
This function interpolates the mean values on a grid of points (\code{DNO}) for a spatiotemporal
nonhomogeneous Poisson model with Goel intensity. The function utilizes the MCMC outputs
from \code{STModelGoelMCMC} and applies a Gaussian process-based interpolation.
}
