% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_mean_surfaceMUSASA.R
\name{compute_mean_surfaceMUSASA}
\alias{compute_mean_surfaceMUSASA}
\title{Compute Interpolated Mean Surface for Spatiotemporal Model}
\usage{
compute_mean_surfaceMUSASA(results, sites, X, DNO, CovXNO, tau)
}
\arguments{
\item{results}{A list containing the output from \code{STModelMusaOkumotoMCMCSA}, including:
\itemize{
\item \code{MW}: Samples for parameter W.
\item \code{MWT}: Acceptance indicators for parameter W.
\item \code{Malpha}: Samples for parameter alpha.
\item \code{MalphaT}: Acceptance indicators for parameter alpha.
\item \code{Mb}: Samples for parameter phi.
\item \code{MbT}: Acceptance indicators for phi.
\item \code{Mv}: Samples for parameter sigma^2.
\item \code{MPsi}: Samples for parameter Psi.
\item \code{Mdelta} : Samples of parameter delta obtained during the MCMC procedure (iteration - burnin).
\item \code{Mtheta} : Samples of parameter theta obtained during the MCMC procedure (iteration - burnin).
\item \code{Mf} : Samples of parameter f obtained during the MCMC procedure (iteration - burnin).
}}

\item{sites}{A matrix with geographic coordinates of the monitoring stations.}

\item{X}{Covariates for the scale parameter of the Weibull intensity.}

\item{DNO}{A grid of points where interpolation is to be performed.}

\item{CovXNO}{Covariates for the scale parameter at the grid points.}

\item{tau}{A vector of temporal points for which the mean surface is computed.}
}
\value{
A list: (\code{Surface}) containing the interpolated mean values at the grid points
for each temporal point in \code{tau}. The first column contains the mean values at the initial
time step, and subsequent columns contain the differences between consecutive time steps.
}
\description{
This function interpolates the mean values on a grid of points (\code{DNO}) for a spatiotemporal
nonhomogeneous Poisson model with Weibull intensity. The function utilizes the MCMC outputs
from \code{STModelMusaOkMCMCSA} and applies a Gaussian process-based interpolation.
}
