% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_mean_surfaceWEIBULLSA.R
\name{compute_mean_surfaceWEIBULLSA}
\alias{compute_mean_surfaceWEIBULLSA}
\title{Compute Interpolated Mean Surface for Spatiotemporal Model}
\usage{
compute_mean_surfaceWEIBULLSA(resultsSA, sites, DNO, CovXNO, CovZNO, tau)
}
\arguments{
\item{resultsSA}{A list containing the output from \code{STModelWeibullMCMCSA}, including:
\itemize{
\item \code{MW}: Samples for parameter W.
\item \code{MWT}: Acceptance indicators for parameter W.
\item \code{MMj}: Samples for parameter M.
\item \code{MMT}: Acceptance indicators for parameter M.
\item \code{Mvw}: Samples for parameter sigma^2_w.
\item \code{Mvm}: Samples for parameter sigma^2_m.
\item \code{MBeta}: Samples for parameter Beta.
\item \code{Mbw}: Samples for parameter phi_w.
\item \code{MbwT}: Acceptance indicators for phi_w.
\item \code{Mbm}: Samples for parameter phi_m.
\item \code{MbmT}: Acceptance indicators for phi_m.
\item \code{MPsi}: Samples for parameter Psi.
\item \code{Mdelta} : Samples of parameter delta obtained during the MCMC procedure (iteration - burnin).
\item \code{MdeltaT} : A binary vector indicating whether each proposed value of delta was accepted (1) or rejected (0). Used to compute the acceptance rate for delta.
\item \code{Mtheta} : Samples of parameter theta obtained during the MCMC procedure (iteration - burnin).
\item \code{MthetaT} : A binary vector indicating whether each proposed value of delta was accepted (1) or rejected (0). Used to compute the acceptance rate for theta.
\item \code{Mf} : Samples of parameter f obtained during the MCMC procedure (iteration - burnin).
\item \code{MfT} : A binary vector indicating whether each proposed value of delta was accepted (1) or rejected (0). Used to compute the acceptance rate for f.
}}

\item{sites}{A matrix with geographic coordinates of the monitoring stations.}

\item{DNO}{A grid of points where interpolation is to be performed.}

\item{CovXNO}{Covariates for the scale parameter at the grid points.}

\item{CovZNO}{Covariates for the shape parameter at the grid points.}

\item{tau}{A vector of temporal points for which the mean surface is computed.}
}
\value{
A list: (\code{Surface}) containing the interpolated mean values at the grid points
for each temporal point in \code{tau}. The first column contains the mean values at the initial
time step, and subsequent columns contain the differences between consecutive time steps.
}
\description{
This function interpolates the mean values on a grid of points (\code{DNO}) for a spatiotemporal
nonhomogeneous Poisson model with Weibull intensity with a seasonal component.. The function utilizes the MCMC outputs
from \code{STModelWeibullMCMCSA} and applies a Gaussian process-based interpolation.
}
