% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_event_times.R
\name{identify_event_times}
\alias{identify_event_times}
\title{Identify Event Times Based on Threshold}
\usage{
identify_event_times(series, threshold)
}
\arguments{
\item{series}{A matrix of dimensions Txn, where each column represents the time series of a random variable from a monitoring station.}

\item{threshold}{A numeric value representing the user-defined threshold to identify times of interest.}
}
\value{
It returns a matrix where each column corresponds to a location, and each row contains the times of occurrence for that location. The matrix dimensions are m x n, where:
\itemize{
\item \code{m}: Maximum number of occurrences across all locations.
\item \code{n}: Number of locations.
}
}
\description{
This function identifies the times at which the values in each column of a matrix exceed a user-defined threshold.
}
\examples{
data(series)
Times <- identify_event_times(series,20)

}
