% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_mean_weibull.R
\name{interpolate_mean_weibull}
\alias{interpolate_mean_weibull}
\title{Interpolate Accumulated Mean Values at an Unobserved Location}
\usage{
interpolate_mean_weibull(results, data, sites, Xw, Xm, Sites1, gridt, Xwr, Xmr)
}
\arguments{
\item{results}{A list containing the output from \code{STModelWeibullMCMC}, including:
\itemize{
\item \code{MW}: Samples for parameter W.
\item \code{MMj}: Samples for parameter M.
\item \code{Mvw}: Samples for variance sigma^2_w.
\item \code{Mvm}: Samples for variance sigma^2_m.
\item \code{Mbw}: Samples for spatial decay phi_w.
\item \code{Mbm}: Samples for spatial decay phi_m.
\item \code{MPsi}: Samples for regression coefficients Psi.
\item \code{MBeta}: Samples for regression coefficients Beta.
}}

\item{data}{A matrix of occurrence times for the event of interest. Each column corresponds to a monitoring station.}

\item{sites}{A matrix of geographic coordinates where the process was observed nx2.}

\item{Xw}{Covariates associated with the scale parameter at the location.}

\item{Xm}{Covariates associated with the shape parameter at the location.}

\item{Sites1}{A vector representing the geographic coordinates of the location where interpolation is to be performed mx2.}

\item{gridt}{A vector of times at which the accumulated mean is to be estimated.}

\item{Xwr}{Covariates associated with the scale parameter at the location where interpolation is to be performed.}

\item{Xmr}{Covariates associated with the shape parameter at the location where interpolation is to be performed.}
}
\value{
A matrix (\code{MatMean}) containing the interpolated accumulated mean values at the times specified in \code{gridt}.
}
\description{
This function interpolates the accumulated mean values for a location where no data was observed,
using outputs from the STModelWeibullMCMC model. The interpolation is performed over a vector of times.
}
