% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_meanSA.R
\name{interpolate_meanSA}
\alias{interpolate_meanSA}
\title{Interpolate Accumulated Mean Values at an Unobserved Location}
\usage{
interpolate_meanSA(resultsSA, data, sites, Xm, Xw, Sites1, gridt, Xmr, Xwr)
}
\arguments{
\item{resultsSA}{A list containing the output from \code{STModelWeibullMCMCSA}, including:
\itemize{
\item \code{MW}: Samples for parameter W.
\item \code{MMj}: Samples for parameter M.
\item \code{Mvw}: Samples for variance sigma^2_w.
\item \code{Mvm}: Samples for variance sigma^2_m.
\item \code{Mbw}: Samples for spatial decay phi_w.
\item \code{Mbm}: Samples for spatial decay phi_m.
\item \code{MPsi}: Samples for regression coefficients Psi.
\item \code{MBeta}: Samples for regression coefficients Beta.
}}

\item{data}{A matrix of occurrence times for the event of interest. Each column corresponds to a monitoring station.}

\item{sites}{A matrix of geographic coordinates where the process was observed.}

\item{Xm}{aaaaaa}

\item{Xw}{aaaa}

\item{Sites1}{A vector representing the geographic coordinates of the location where interpolation is to be performed.}

\item{gridt}{A vector of times at which the accumulated mean is to be estimated.}

\item{Xmr}{Covariates associated with the shape parameter at the location.}

\item{Xwr}{Covariates associated with the scale parameter at the location.}
}
\value{
A matrix (\code{MatMean}) containing the interpolated accumulated mean values at the times specified in \code{gridt}.
}
\description{
This function interpolates the accumulated mean values for a location where no data was observed,
using outputs from the STModelWeibullMCMC model. The interpolation is performed over a vector of times.
}
