% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STModelGoelMCMC.R
\name{STModelGoelMCMC}
\alias{STModelGoelMCMC}
\title{Title}
\usage{
STModelGoelMCMC(
  data,
  sites,
  X = cbind(as.matrix(rep(1, ncol(data))), as.matrix(sites)),
  Z = X,
  M = Z,
  prior = list(c3 = (-2 * log(0.05)/max(dist(sites))) * 0.1, d3 = 0.1, BB1 = diag(100,
    ncol(Z)), AA1 = as.matrix(rep(0, ncol(Z))), aa1 = 2.01, bb1 = 1.005, V = diag(100,
    ncol(X)), MM1 = as.matrix(rep(0, ncol(X))), SU1 = 0.01, SU2 = 0.01, SU3 = 28.01968,
    SU5 = 0.001118574, SU6 = 0.0553372),
  iter = 410000,
  bar = 4e+05
)
}
\arguments{
\item{data}{aaaa}

\item{sites}{aaaa}

\item{X}{aaaa}

\item{Z}{aaaa}

\item{M}{aaaa}

\item{prior}{aaaa}

\item{iter}{aaaa}

\item{bar}{aaaa}
}
\value{
aaaa
}
\description{
Title
}
