% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STModelGoelMCMCSA.R
\name{STModelGoelMCMCSA}
\alias{STModelGoelMCMCSA}
\title{Title}
\usage{
STModelGoelMCMCSA(
  data,
  sites,
  X = cbind(as.matrix(rep(1, ncol(data))), sites),
  Z = cbind(as.matrix(rep(1, ncol(data))), sites),
  Fj = Z,
  prior = list(Psi = as.matrix(rep(0, ncol(X))), BB1 = diag(100, ncol(Z)), AA1 =
    as.matrix(rep(0, ncol(Z))), BB2 = diag(100, ncol(Z)), AA2 = as.matrix(rep(0,
    ncol(Z))), A = as.matrix(rep(0, ncol(Z))), B = diag(100, ncol(Z)), lgama =
    as.matrix(rep(0.01, ncol(Z))), leta = as.matrix(rep(1, ncol(Z))), aa1 = 2.01, bb1 =
    1.005, c3 = (-2 * log(0.05)/max(dist(sites))) * 0.1, d3 = 0.1),
  iteration,
  burnin
)
}
\arguments{
\item{data}{aaa}

\item{sites}{aaa}

\item{X}{aaa}

\item{Z}{aaa}

\item{Fj}{aaa}

\item{prior}{aaa}

\item{iteration}{aaa}

\item{burnin}{aaa}
}
\value{
aaaa
}
\description{
Title
}
