% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STModelMusaOkumotoMCMC.R
\name{STModelMusaOkumotoMCMC}
\alias{STModelMusaOkumotoMCMC}
\title{Title}
\usage{
STModelMusaOkumotoMCMC(
  data,
  sites,
  prior = list(aa1 = 0.001, bb1 = 0.001, c2 = 1e-05, d2 = 0.001, c4 = 2.01, d4 = 1.005,
    aa2 = 2.01, bb2 = 1.005, c3 = (-2 * log(0.05)/max(dist(sites))), d3 = 0.1),
  iteration,
  burnin
)
}
\arguments{
\item{data}{aaaa}

\item{sites}{aaaa}

\item{prior}{aaaa}

\item{iteration}{aaaa}

\item{burnin}{aaaaaa}
}
\value{
aaaaa
}
\description{
Title
}
