% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STModelWeibullMCMC.R
\name{STModelWeibullMCMC}
\alias{STModelWeibullMCMC}
\title{Algoritmo MCMC Poisson Process}
\usage{
STModelWeibullMCMC(
  data,
  sites,
  X = cbind(as.matrix(rep(1, ncol(data))), as.matrix(sites)),
  Z = X,
  prior = list(aa1 = 0.001, bb1 = 0.001, aa2 = 2.01, bb2 = 1.005, c3 = (-2 *
    log(0.05)/max(dist(sites))) * 0.1, d3 = 0.1, SU1 = 1e-06, SU2 = 1e-06, SU3 = 100, SU4
    = 100, Psi = as.matrix(rep(0, ncol(X))), A1 = as.matrix(rep(0, ncol(X))), B1 =
    diag(100, ncol(X)), Beta = as.matrix(rep(0, ncol(Z))), A = as.matrix(rep(0,
    ncol(Z))), B = diag(100, ncol(Z))),
  iteration,
  burnin
)
}
\arguments{
\item{data}{Matriz de dados simulados de um processo de Poisson}

\item{sites}{Malha de pontos}

\item{X}{aaa}

\item{Z}{aaa}

\item{prior}{description}

\item{iteration}{aaa}

\item{burnin}{aaaa}
}
\description{
Esta é a função para realizar o algoritmo Monte Carlo via Cadeias de Markov para os parâmetros de um processo de Poisson.
}
