% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_goel_model.R
\name{fit_goel_model}
\alias{fit_goel_model}
\title{Fit and Plot Nonhomogeneous Poisson Goel Model}
\usage{
fit_goel_model(data, results, Z, M, l)
}
\arguments{
\item{data}{A matrix of occurrence times for the event of interest. Each column corresponds to a monitoring station.}

\item{results}{A list containing the output from \code{STModelWeibullMCMC}, which includes:
\itemize{
\item \code{MW}: Samples for parameter W.
\item \code{Mbeta}: Samples for parameter Beta.
\item \code{Malpha}: Samples for parameter Alpha.
}}

\item{Z}{A matrix of covariates associated with the parameter beta.}

\item{M}{A matrix of covariates associated with the parameter alpha. Default is the same as \code{Z}.}

\item{l}{An integer specifying the index of the monitoring station to analyze.}
}
\value{
A list containing:
\describe{
\item{retemp}{The occurrence times of the events of interest.}
\item{medy1}{The estimated cumulative mean function at the event times.}
\item{mtnp}{The observed cumulative mean function at the event times.}
}
}
\description{
This function fits a nonhomogeneous Poisson Goel model to data from a specified monitoring station,
generates a plot of the observed and estimated cumulative mean functions, and returns relevant metrics.
}
