% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_musa_modelSA.R
\name{fit_musa_modelSA}
\alias{fit_musa_modelSA}
\title{Fit and Plot Nonhomogeneous Poisson Musa Okumoto Model}
\usage{
fit_musa_modelSA(data, results, l)
}
\arguments{
\item{data}{A matrix of occurrence times for the event of interest. Each column corresponds to a monitoring station.}

\item{results}{A list containing the output from \code{STModelMusaOkumotoMCMCSA}, which includes:
\itemize{
\item \code{MW}: Samples for parameter W.
\item \code{Malpha}: Samples for parameter Alpha.
\item \code{Mdelta} : Samples of parameter delta obtained during the MCMC procedure (iteration - burnin).
\item \code{Mtheta} : Samples of parameter theta obtained during the MCMC procedure (iteration - burnin).
\item \code{Mf} : Samples of parameter f obtained during the MCMC procedure (iteration - burnin).
}}

\item{l}{An integer specifying the index of the monitoring station to analyze.}
}
\value{
A list containing:
\describe{
\item{retemp}{The occurrence times of the events of interest.}
\item{medy1}{The estimated cumulative mean function at the event times.}
\item{mtnp}{The observed cumulative mean function at the event times.}
}
}
\description{
This function fits a nonhomogeneous Poisson Musa Okumoto model with a seasonal component. to data from a specified monitoring station,
generates a plot of the observed and estimated cumulative mean functions, and returns relevant metrics.
}
