% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_event_times.R
\name{identify_event_times}
\alias{identify_event_times}
\title{Identify Event Times Based on Threshold}
\usage{
identify_event_times(series, threshold)
}
\arguments{
\item{series}{A matrix of dimensions Txn, where each column represents the time series of a random variable from a monitoring station.}

\item{threshold}{A numeric value representing the user-defined threshold to identify times of interest.}
}
\value{
A matrix where each column contains the times (indices) at which the values of the respective column in \code{series} exceeded the threshold. Rows are padded with NA where no event occurred.
}
\description{
This function identifies the times at which the values in each column of a matrix exceed a user-defined threshold.
}
