% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STModelMusaOkumotoMCMCSA.R
\name{STModelMusaOkumotoMCMCSA}
\alias{STModelMusaOkumotoMCMCSA}
\title{Title}
\usage{
STModelMusaOkumotoMCMCSA(
  data,
  sites,
  X = cbind(as.matrix(rep(1, ncol(data))), (1/100) * sites),
  prior = list(Psi = as.matrix(rep(0, ncol(X))), A = as.matrix(rep(0, ncol(X))), B =
    diag(100, ncol(X)), V = diag(100, ncol(X)), M = as.matrix(rep(0, ncol(X))), c2 =
    1e-05, d2 = 0.001, c4 = 2.01, d4 = 1.005, aa2 = 2.01, bb2 = 1.005, aa1 = 2.01, bb1 =
    1.005, c3 = (-2 * log(0.05)/max(dist(sites))) * 0.1, d3 = 0.1),
  iteration,
  burnin
)
}
\arguments{
\item{data}{aaaa}

\item{sites}{aaaa}

\item{X}{aaaa}

\item{prior}{aaaa}

\item{iteration}{aaaa}

\item{burnin}{aaaa}
}
\value{
aaaa
}
\description{
Title
}
